Create separate packages for the different aspects of the custom data format.
Currently just separating the BallData class from the CarData and CarOrientation classes, away from the DataPacket class.
