+++ src/main/java/rlbotexample/SampleBot.java
@@ -10,8 +10,8 @@ import rlbot.flat.QuickChatSelection;
 import rlbot.manager.BotLoopRenderer;
 import rlbot.render.Renderer;
 import rlbotexample.boost.BoostManager;
-import rlbotexample.input.CarData;
 import rlbotexample.input.DataPacket;
+import rlbotexample.input.car.CarData;
 import rlbotexample.output.ControlsOutput;
 import rlbotexample.prediction.BallPredictionHelper;
 import rlbotexample.vector.Vector2;
