+++ src/main/java/rlbotexample/input/car/CarData.java
@@ -1,4 +1,4 @@
-package rlbotexample.input;
+package rlbotexample.input.car;
 
 
 import rlbotexample.vector.Vector3;
