+++ src/main/java/rlbotexample/input/car/CarOrientation.java
@@ -1,4 +1,4 @@
-package rlbotexample.input;
+package rlbotexample.input.car;
 
 
 import rlbot.flat.PlayerInfo;
